api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pip-audit
requests
ruff
python-multipart 